Refactor to new layout
