+++ include/NovelRT/Ecs/Input/Ecs.Input.h
@@ -8,7 +8,7 @@
 #error NovelRT does not support including types explicitly by default. Please include Ecs.h instead for the Ecs namespace subset.
 #endif
 
-#include "../../Input/Input.h"
+#include <NovelRT/Input/Input.hpp>
 #include <map>
 
 namespace NovelRT::Ecs::Input
