+++ include/NovelRT/NovelRT.h
@@ -114,8 +114,8 @@
   #include <NovelRT/Threading/Threading.h>
   #include <NovelRT/Windowing/Windowing.h>
   #include <NovelRT/Windowing/Glfw/Windowing.Glfw.h>
-  #include <NovelRT/Input/Input.h>
-  #include <NovelRT/Input/Glfw/Input.Glfw.h>
+  #include <NovelRT/Input/Input.hpp>
+  #include <NovelRT/Input/Glfw/Input.Glfw.hpp>
 
   // Plugin Management types
   #include <NovelRT/PluginManagement/PluginManagement.h>
