+++ input/Glfw/GlfwInputPluginProvider.cpp
@@ -1,7 +1,7 @@
 // Copyright © Matt Jones and Contributors. Licensed under the MIT Licence (MIT). See LICENCE.md in the repository root
 // for more information.
 
-#include <NovelRT/Input/Glfw/Input.Glfw.h>
+#include <NovelRT/Input/Glfw/Input.Glfw.hpp>
 
 namespace NovelRT::Input::Glfw
 {
