+++ input/include/NovelRT/Input/IInputDevice.hpp
