+++ input/include/NovelRT/Input/KeyStateFrameChangeLog.hpp
