+++ src/NovelRT/PluginManagement/TemporaryFnPtrs.cpp
@@ -2,7 +2,7 @@
 // for more information.
 
 #include <NovelRT/Graphics/Vulkan/Graphics.Vulkan.hpp>
-#include <NovelRT/Input/Glfw/Input.Glfw.h>
+#include <NovelRT/Input/Glfw/Input.Glfw.hpp>
 #include <NovelRT/PluginManagement/PluginManagement.h>
 #include <NovelRT/PluginManagement/TemporaryFnPtrs.h>
 #include <NovelRT/ResourceManagement/Desktop/ResourceManagement.Desktop.h>
